Refactor: mongodb split
