+++ backend/package.json
@@ -29,12 +29,12 @@
     "@nestjs/passport": "^10.0.3",
     "@nestjs/platform-express": "^10.3.0",
     "@nestjs/serve-static": "^4.0.0",
-    "@prisma/client": "^5.8.0",
+    "@prisma/client": "^5.8.1",
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
-    "dotenv": "^16.3.1",
+    "dotenv": "^16.3.2",
     "express-session": "^1.17.3",
     "passport": "^0.6.0",
     "passport-local": "^1.0.0",
@@ -51,21 +51,21 @@
     "@types/express-session": "^1.17.10",
     "@types/jest": "^29.5.11",
     "@types/multer": "^1.4.11",
-    "@types/node": "^20.11.0",
+    "@types/node": "^20.11.5",
     "@types/passport-local": "^1.0.38",
     "@types/session-file-store": "^1.2.5",
     "@types/supertest": "^2.0.16",
-    "@typescript-eslint/eslint-plugin": "^6.18.1",
-    "@typescript-eslint/parser": "^6.18.1",
+    "@typescript-eslint/eslint-plugin": "^6.19.0",
+    "@typescript-eslint/parser": "^6.19.0",
     "dotenv-cli": "^7.3.0",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-prettier": "^5.1.3",
     "jest": "^29.7.0",
-    "prettier": "^3.2.1",
-    "prisma": "^5.8.0",
+    "prettier": "^3.2.4",
+    "prisma": "^5.8.1",
     "source-map-support": "^0.5.21",
-    "supertest": "^6.3.3",
+    "supertest": "^6.3.4",
     "ts-jest": "^29.1.1",
     "ts-loader": "^9.5.1",
     "ts-node": "^10.9.2",
