+++ backend/src/modules/mongodb/__mocks__/mongodb.postimage.service.ts
@@ -1,0 +1,6 @@
+import { imageStub } from './stubs/mongodb.stub';
+
+export const MongodbPostImageService = jest.fn().mockReturnValue({
+  getImage: jest.fn().mockReturnValue(imageStub()),
+  saveImage: jest.fn().mockReturnValue(imageStub()),
+});
