+++ backend/src/modules/mongodb/__mocks__/mongodb.postkey.service.ts
@@ -1,0 +1,5 @@
+import { postStub } from './stubs/mongodb.stub';
+
+export const MongodbPostKeyService = jest.fn().mockReturnValue({
+  getPostKey: jest.fn().mockReturnValue(1),
+});
