+++ backend/src/modules/reservation/reservation.controller.spec.ts
@@ -1,16 +1,17 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ReservationController } from './reservation.controller';
 import { ReservationService } from './reservation.service';
-import { MongodbService } from '../mongodb/mongodb.service';
 import { PrismaService } from '../prisma/prisma.service';
+import { MongodbModule } from '../mongodb/mongodb.module';
 
 describe('ReservationController', () => {
   let controller: ReservationController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
+      imports: [MongodbModule],
       controllers: [ReservationController],
-      providers: [ReservationService, MongodbService, PrismaService],
+      providers: [ReservationService, PrismaService],
     }).compile();
 
     controller = module.get<ReservationController>(ReservationController);
